requirements.txt
aiogram
sqlalchemy
aiosqlite
python-dotenv
yookassa
g4f